rename schema files with camel case
