+++ registry/modules/authBasic/router/auth.ts
@@ -1,6 +1,6 @@
 import express from "express";
 
-import { basicAuthValidator } from "@/schemaValidators/auth-basic.zod.js";
+import { basicAuthValidator } from "@/schemaValidators/authBasic.zod.js";
 
 import { createAuthBasicController } from "@/modules/authBasic/controllers/authBasic.js";
 import { createUserRepository } from "@/repositories/user.postgres.js";
