+++ registry/schemaValidators/authBasic.interface.ts
